Fix photometric calculations
Prior to this commit, the flux was computed simply as the sum of
the active pixels in each frame. Unfortunately, this approach is
incorrect because you must know the maximum flux of all frames to
properly de-normalize the photometry. This commit exposes the maximum
flux from each frame so photometry is correctly computed.